[modify] 스터디룸 페이지 파일 위치 변경
