+++ src/components/studyRoom/diary/GroupDiary.tsx
@@ -1,0 +1,5 @@
+const GroupDiary = () => {
+  return <div>GroupDiary</div>;
+};
+
+export default GroupDiary;
