+++ src/components/studyRoom/home/Chats/ChatList.tsx
@@ -1,0 +1,6 @@
+const ChatList = () => {
+  // 채팅 조회
+  return <div className="flex-1 bg-[#F7F6F6]"></div>;
+};
+
+export default ChatList;
